feat(Colors): Changed lighter and darker naming convention from numbers to words
BREAKING CHANGE: Color variable names have changed. Previously dark and light variations were numbered 0 and 2, with the mid color numbered 1. For example

```css
--color-ui-primary-0 -> --color-ui-primary--dark
--color-ui-primary-1 -> --color-ui-primary
--color-ui-primary-2 -> --color-ui-primary--light
```

With the exception of the `--color-ui-dark` color, which was origianlly numbered incorrectly, this has been corrected but it maps to the opposite (if you're doing a bulk find and replace)

```css
--color-ui-dark-0 -> --color-ui-dark--light
--color-ui-dark-1 -> --color-ui-dark
--color-ui-dark-2 -> --color-ui-dark--dark
```