feat(Theming): Support and combination of light and dark themes
BREAKING CHANGE:

Axiom has now changed to not preprocess css variables by default. This was needed to support nested theming and as a result IE11 is no longer supported out of the box, however using PostCSS and the provided CSS variables partial IE11 support can be added but will not have nested theme support. Please see the IE11 support documentation on how to set it up.

For non IE11 projects follow the following steps to migrate over.

1) Remove all css variables imports from your css files as these are no longer needed, for example...

```css
- @import 'bw-axiom/lib/materials/sizing';
```

There is the exception if you are using the custom media breakpoints, in which case you will need to include these into your css with...

```css
+ @import 'bw-axiom/lib/materials/breakpoints.css';
```

2) Add the class `ax-theme--light` or `ax-theme--dark` to the *<html>* element.
3) <TextInput> when `theme` was used before, the prop should still remain on the component but the prop `style="overlay"` should be added