fix(ContextMenuItem): mouse events do not fire in firefox
A [bug](https://bugzilla.mozilla.org/show_bug.cgi?id=1084196) in firefox
cause mouse events to not fire below a button element causing the
EllipseTooltip to not work inside a DropdownContextMenuItem in Firefox.

Generally it makes sense to use a button for a ContextMenuItem as long
as it is a simple entry you can click.

But there are at least two use-cases that allow some sort of interaction
inside the ContextMenuItem:

1. checkbox to activate ContextMenuItem
2. showing a tooltip on a truncated named (using the EllipseTooltip)

Though the definition of the button does [permit interactive
content](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button).

As there are use-cases where there is interaction in the ContextMenuItem are it
is reasonable to change the element of ContextMenuItem to a div but set
the ARIA role to `MenuItem` instead to fix that bug.