feat: default duration picker selected unit to days if valid
if days is not excludes then the selected unit by default will be days
days is often the most supported unit of time. This is particularly
useful in Vizia where we want to update a number of components at once
but not all support mins/hours etc. only day. We do not want to exclude
these other options as they are still valid for a subset of the components